+++ annotlib/ANNOTmenu.c
@@ -777,7 +777,7 @@ LRESULT CALLBACK WIN_AnnotFilterDlgProc (ThotWindow hwnDlg, UINT msg,
   ----------------------------------------------------------------------*/
 static void AnnotFilterCallbackDialog (int ref, int typedata, char * data)
 {
-  int val;
+  intptr_t val;
 
   if (ref == -1)
     {
@@ -787,7 +787,7 @@ static void AnnotFilterCallbackDialog (int ref, int typedata, char * data)
   else
     {
       /* has the user changed the options? */
-      val = (int) data;
+      val = (intptr_t) data;
       switch (ref - AnnotFilterBase)
         {
         case AnnotFilterMenu:
